playwright

+++ backend/src/services/posts.js
@@ -1,7 +1,8 @@
 import { Post } from '../db/models/post.js'
+import { User } from '../db/models/user.js'
 
-export async function createPost({ title, author, contents, tags }) {
-  const post = new Post({ title, author, contents, tags })
+export async function createPost(userId, { title, contents, tags }) {
+  const post = new Post({ title, author: userId, contents, tags })
   return await post.save()
 }
 
@@ -19,8 +20,10 @@ async function listPosts(
 export async function listAllPosts(options) {
   return await listPosts({}, options)
 }
-export async function listPostsByAuthor(author, options) {
-  return await listPosts({ author }, options)
+export async function listPostsByAuthor(authorUsername, options) {
+  const user = await User.findOne({ ussername: authorUsername })
+  if (!user) return []
+  return await listPosts({ author: user._id }, options)
 }
 
 export async function listPostsByTag(tags, options) {
@@ -31,14 +34,20 @@ export async function listPostsByTag(tags, options) {
 export async function getPostById(postId) {
   return await Post.findById(postId)
 }
-export async function updatePost(postId, { title, author,
-  contents, tags }) {
+export async function updatePost(userId, postId, { title, contents, tags }) {
   return await Post.findOneAndUpdate(
-    { _id: postId },
-    { $set: { title, author, contents, tags } },
+    { _id: postId, author: userId },
+    { $set: { title, contents, tags } },
     { new: true },
   )
 }
-export async function deletePost(postId) {
-  return await Post.deleteOne({ _id: postId })
+export async function deletePost(userId, postId) {
+  return await Post.deleteOne({ _id: postId, author: userId })
 }
+
+/* 
+We could
+still extend this code to first fetch the post with the given ID, check if it exists (if not, return
+a 404 Not Found error), and if it does exist, verify that the author is the currently logged-in
+user (if not, return a 403 Forbidden error).
+*/
